agent-orchestrator/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0
groq>=0.4.0

# Database
asyncpg>=0.29.0
redis>=5.0.0
qdrant-client>=1.7.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
